// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface NavigationMenuSliceDefaultItem {
  /**
   * ChildText field in *NavigationMenu → Items*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: navigation_menu.items[].childtext
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  childtext: prismic.KeyTextField;

  /**
   * ChildLink field in *NavigationMenu → Items*
   *
   * - **Field Type**: Link
   * - **Placeholder**: *None*
   * - **API ID Path**: navigation_menu.items[].childlink
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  childlink: prismic.LinkField;
}

/**
 * Default variation for NavigationMenu Slice
 *
 * - **API ID**: `default`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type NavigationMenuSliceDefault = prismic.SharedSliceVariation<
  "default",
  Simplify<NavigationMenuSliceDefaultPrimary>,
  Simplify<NavigationMenuSliceDefaultItem>
>;
// ...
export type NavigationMenuSliceSingleItem = prismic.SharedSliceVariation<
  "singleItem",
  Simplify<NavigationMenuSliceSingleItemPrimary>,
  never
>;
// ...
export interface NavigationMenuSliceSeriesListingItem {
  /**
   * series field in *NavigationMenu → Items*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: navigation_menu.items[].series
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  series: prismic.ContentRelationshipField<"series">;
}

/**
 * SeriesListing variation for NavigationMenu Slice
 *
 * - **API ID**: `seriesListing`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type NavigationMenuSliceSeriesListing = prismic.SharedSliceVariation<
  "seriesListing",
  Simplify<NavigationMenuSliceSeriesListingPrimary>,
  Simplify<NavigationMenuSliceSeriesListingItem>
>;
// ...
export interface NavigationMenuSliceEventListingItem {
  /**
   * event field in *NavigationMenu → Items*
   *
   * - **Field Type**: Content Relationship
   * - **Placeholder**: *None*
   * - **API ID Path**: navigation_menu.items[].event
   * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
   */
  event: prismic.ContentRelationshipField<"event">;
}

/**
 * EventListing variation for NavigationMenu Slice
 *
 * - **API ID**: `eventListing`
 * - **Description**: Default
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type NavigationMenuSliceEventListing = prismic.SharedSliceVariation<
  "eventListing",
  Simplify<NavigationMenuSliceEventListingPrimary>,
  Simplify<NavigationMenuSliceEventListingItem>
>;

/**
 * Slice variation for *NavigationMenu*
 */
type NavigationMenuSliceVariation =
  | NavigationMenuSliceDefault
  | NavigationMenuSliceSingleItem
  | NavigationMenuSliceSeriesListing
  | NavigationMenuSliceEventListing;

/**
 * NavigationMenu Shared Slice
 *
 * - **API ID**: `navigation_menu`
 * - **Description**: NavigationMenu
 * - **Documentation**: https://prismic.io/docs/slice
 */
export type NavigationMenuSlice = prismic.SharedSlice<
  "navigation_menu",
  NavigationMenuSliceVariation
>;